+++ source/renderer/app/components/staking/stake-pools/StakePools.js
@@ -94,6 +94,7 @@ type State = {
   search: string,
   selectedList?: ?string,
   isGridView: boolean,
+  isGridRewardsView: boolean,
   isListView: boolean,
   isTableHeaderHovered: boolean,
 };
@@ -102,6 +103,7 @@ const initialState = {
   search: '',
   selectedList: null,
   isGridView: true,
+  isGridRewardsView: false,
   isListView: false,
   isTableHeaderHovered: false,
 };
@@ -123,12 +125,22 @@ export default class StakePools extends Component<Props, State> {
   handleGridView = () =>
     this.setState({
       isGridView: true,
+      isGridRewardsView: false,
       isListView: false,
     });
+
+  handleGridRewardsView = () => {
+    this.setState({
+      isGridView: false,
+      isGridRewardsView: true,
+      isListView: false,
+    });
+  };
 
   handleListView = () =>
     this.setState({
       isGridView: false,
+      isGridRewardsView: false,
       isListView: true,
     });
 
@@ -172,6 +184,7 @@ export default class StakePools extends Component<Props, State> {
       selectedList,
       isListView,
       isGridView,
+      isGridRewardsView,
       isTableHeaderHovered,
     } = this.state;
 
@@ -263,11 +276,12 @@ export default class StakePools extends Component<Props, State> {
               onSearch={this.handleSearch}
               onClearSearch={this.handleClearSearch}
               onGridView={this.handleGridView}
+              onGridRewardsView={this.handleGridRewardsView}
               onListView={this.handleListView}
               isListView={isListView}
               isGridView={isGridView}
+              isGridRewardsView={isGridRewardsView}
               smashServer={smashServer}
-              isClearTooltipOpeningDownward
             />
             {stakePoolsDelegatingList.length > 0 && (
               <Fragment>
@@ -285,6 +299,7 @@ export default class StakePools extends Component<Props, State> {
                   containerClassName="StakingWithNavigation_page"
                   onSelect={this.onDelegate}
                   numberOfRankedStakePools={numberOfRankedStakePools}
+                  isGridRewardsView={isGridRewardsView}
                   showWithSelectButton
                 />
               </Fragment>
@@ -321,7 +336,7 @@ export default class StakePools extends Component<Props, State> {
                 />
               </Fragment>
             )}
-            {isGridView && (
+            {(isGridView || isGridRewardsView) && (
               <Fragment>
                 <h2>
                   <span className={styles.leftContent}>
@@ -344,6 +359,7 @@ export default class StakePools extends Component<Props, State> {
                   containerClassName="StakingWithNavigation_page"
                   onSelect={this.onDelegate}
                   numberOfRankedStakePools={numberOfRankedStakePools}
+                  isGridRewardsView={isGridRewardsView}
                 />
               </Fragment>
             )}