+++ source/renderer/app/components/staking/widgets/PoolPopOver.js
@@ -30,6 +30,7 @@ export function PoolPopOver(props: {
   stakePool: StakePool,
   containerClassName: string,
   numberOfRankedStakePools: number,
+  isGridRewardsView?: boolean,
 }) {
   // Track hover state manually to optimize performance by lazy init pop overs
   const [isHovered, setIsHovered] = useState(false);
@@ -88,6 +89,7 @@ export function PoolPopOver(props: {
               }}
               showWithSelectButton={props.showWithSelectButton}
               stakePool={props.stakePool}
+              isGridRewardsView={props.isGridRewardsView}
             />
           }
         />