+++ source/renderer/app/components/staking/widgets/TooltipPool.js
@@ -164,6 +164,7 @@ type Props = {
   color: string,
   containerClassName: string,
   numberOfRankedStakePools: number,
+  isGridRewardsView?: boolean,
 };
 
 type State = {
@@ -218,7 +219,12 @@ export default class TooltipPool extends Component<Props, State> {
   renderDescriptionFields = () => {
     const { isIncentivizedTestnet } = global;
     const { intl } = this.context;
-    const { currentTheme, stakePool, numberOfRankedStakePools } = this.props;
+    const {
+      currentTheme,
+      stakePool,
+      numberOfRankedStakePools,
+      isGridRewardsView,
+    } = this.props;
     const {
       ranking,
       relativeStake,
@@ -371,9 +377,17 @@ export default class TooltipPool extends Component<Props, State> {
         key: 'potentialRewards',
         value: (
           <div className={styles.defaultColor}>
-            <span className={styles.defaultColorContent}>
-              {formattedWalletAmount(potentialRewards)}
-            </span>
+            {isGridRewardsView &&
+            potentialRewards.isZero &&
+            potentialRewards.isZero() ? (
+              <div className={styles.noDataDash}>
+                <SVGInline svg={noDataDashSmallImage} />
+              </div>
+            ) : (
+              <span className={styles.defaultColorContent}>
+                {formattedWalletAmount(potentialRewards)}
+              </span>
+            )}
           </div>
         ),
       },